Another conversion warning

+++ src/spotprice_cron.cpp
@@ -53,7 +53,7 @@ void SpotpriceCron::main()
         if (!first)
         {
           poll_time = UTCTime().IncrementSecondsCopy(20*60);
-          poll_time.SetMinute((poll_time.GetMinute()/20)*20); //Integer division to round down to 00|20|40
+          poll_time.SetMinute(static_cast<uint8_t>((poll_time.GetMinute()/20)*20)); //Integer division to round down to 00|20|40
           poll_time.SetSecond(0);
           std::this_thread::sleep_until(std::chrono::system_clock::from_time_t(poll_time.AsUTCTimeT()));
         }
